requirements.txt
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.0.1
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2
envparse==0.2.0
fastapi==0.103.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
passlib==1.7.4
psycopg2==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
python-jose==3.3.0